bubble-container 파일 정리
