+++ packages/chat/src/bubble-container/elements.tsx
@@ -1,5 +1,5 @@
 import styled, { css } from 'styled-components'
-import { Text } from '@titicaca/core-elements'
+import { Button, Text } from '@titicaca/core-elements'
 
 export const HiddenElement = styled.div`
   height: 1px;
@@ -51,3 +51,45 @@ export const DeleteButton = styled.button`
   border-left: 1px solid rgba(255, 255, 255, 0.3);
   background-image: url('https://assets.triple.guide/images/btn-message-delete@3x.png');
 `
+
+const ThanksButton = styled(Button)<{ haveMine: boolean }>`
+  display: flex;
+  gap: 3px;
+  height: 20px;
+  align-items: center;
+  background-color: ${({ haveMine }) =>
+    haveMine ? 'var(--color-white)' : 'var(--color-gray50)'};
+  color: ${({ haveMine }) => (haveMine ? '#1DBEB2' : 'var(--color-gray700)')};
+  ${({ haveMine }) => (haveMine ? 'border: 1px solid #1DBEB2;' : '')}
+  padding: 3.5px 6px 4.5px 7px;
+  font-weight: ${({ haveMine }) => (haveMine ? '700' : '500')};
+  font-size: 10px;
+`
+
+const ThanksCount = styled.span`
+  font-size: 10px;
+  line-height: 11px;
+`
+
+export function Thanks({
+  count,
+  haveMine,
+  onClick,
+  ...props
+}: {
+  count: number
+  haveMine: boolean
+  onClick?: () => void
+}) {
+  return (
+    <ThanksButton haveMine={haveMine} onClick={() => onClick?.()} {...props}>
+      <img
+        src="https://assets.triple.guide/images/ic_chat_thumbsup_on.svg"
+        alt="좋아요 아이콘"
+        width={11}
+        height={11}
+      />
+      {count === 0 ? null : <ThanksCount>{count}</ThanksCount>}
+    </ThanksButton>
+  )
+}
